+++ tests/e2e/Services/Functions/FunctionsCustomServerTest.php
@@ -608,7 +608,7 @@ class FunctionsCustomServerTest extends Scope
         $this->assertEquals(0, $execution['body']['statusCode']);
         $this->assertEquals('', $execution['body']['response']);
         $this->assertEquals('', $execution['body']['stderr']);
-        $this->assertEquals(0, $execution['body']['time']);
+        $this->assertEquals(0, $execution['body']['duration']);
 
         sleep(5);
 
@@ -631,7 +631,7 @@ class FunctionsCustomServerTest extends Scope
         $this->assertStringContainsString('8.0', $execution['body']['response']);
         $this->assertStringContainsString('êä', $execution['body']['response']); // tests unknown utf-8 chars
         $this->assertEquals('', $execution['body']['stderr']);
-        $this->assertLessThan(0.500, $execution['body']['time']);
+        $this->assertLessThan(0.500, $execution['body']['duration']);
 
         /**
          * Test for FAILURE
@@ -738,7 +738,7 @@ class FunctionsCustomServerTest extends Scope
             'content-type' => 'application/json',
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
-            'async' => false,
+            // Testing default value, should be 'async' => false
         ]);
 
         $this->assertEquals(201, $execution['headers']['status-code']);
@@ -750,7 +750,7 @@ class FunctionsCustomServerTest extends Scope
         $this->assertStringContainsString('PHP', $execution['body']['response']);
         $this->assertStringContainsString('8.0', $execution['body']['response']);
         $this->assertStringContainsString('êä', $execution['body']['response']); // tests unknown utf-8 chars
-        $this->assertLessThan(0.500, $execution['body']['time']);
+        $this->assertLessThan(0.500, $execution['body']['duration']);
 
         return $data;
     }
@@ -909,9 +909,9 @@ class FunctionsCustomServerTest extends Scope
         $this->assertEquals($executions['body']['executions'][0]['trigger'], 'http');
         $this->assertEquals($executions['body']['executions'][0]['status'], 'failed');
         $this->assertEquals($executions['body']['executions'][0]['statusCode'], 500);
-        $this->assertGreaterThan(2, $executions['body']['executions'][0]['time']);
-        $this->assertLessThan(6, $executions['body']['executions'][0]['time']);
-        $this->assertGreaterThan(4, $executions['body']['executions'][0]['time']);
+        $this->assertGreaterThan(2, $executions['body']['executions'][0]['duration']);
+        $this->assertLessThan(6, $executions['body']['executions'][0]['duration']);
+        $this->assertGreaterThan(4, $executions['body']['executions'][0]['duration']);
         $this->assertEquals($executions['body']['executions'][0]['response'], '');
         $this->assertEquals($executions['body']['executions'][0]['stderr'], 'An internal curl error has occurred within the executor! Error Msg: Operation timed out');
 
@@ -979,6 +979,7 @@ class FunctionsCustomServerTest extends Scope
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
             'data' => 'foobar',
+            'async' => true
         ]);
 
         $executionId = $execution['body']['$id'] ?? '';
@@ -1093,6 +1094,7 @@ class FunctionsCustomServerTest extends Scope
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
             'data' => 'foobar',
+            'async' => true
         ]);
 
         $executionId = $execution['body']['$id'] ?? '';
@@ -1205,6 +1207,7 @@ class FunctionsCustomServerTest extends Scope
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
             'data' => 'foobar',
+            'async' => true
         ]);
 
         $executionId = $execution['body']['$id'] ?? '';
@@ -1318,6 +1321,7 @@ class FunctionsCustomServerTest extends Scope
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
             'data' => 'foobar',
+            'async' => true
         ]);
 
         $executionId = $execution['body']['$id'] ?? '';
@@ -1431,6 +1435,7 @@ class FunctionsCustomServerTest extends Scope
             'x-appwrite-project' => $this->getProject()['$id'],
         ], $this->getHeaders()), [
             'data' => 'foobar',
+            'async' => true
         ]);
 
         $executionId = $execution['body']['$id'] ?? '';